--- app/[lang]/(pages)/biography/page.tsx
@@ -1,5 +1,0 @@
-import React, { ReactElement } from 'react';
-
-export default function Biography(): ReactElement {
-  return <div>Biography</div>;
-}
