--- app/[lang]/(pages)/media-about-us/page.tsx
@@ -1,6 +1,0 @@
-import { Typography } from '@mui/material';
-import React, { ReactElement } from 'react';
-
-export default function MediaAboutUs(): ReactElement {
-  return <Typography variant="h1">Media about us</Typography>;
-}
