+++ app/[lang]/artistry/page.test.tsx
