+++ app/[lang]/biography/page.tsx
@@ -1,0 +1,11 @@
+import React, { ReactElement } from 'react';
+
+import MainLayout from '~/layouts/main-layout/MainLayout';
+
+export default function Biography(): ReactElement {
+  return (
+    <MainLayout withLines>
+      <div>Biography</div>
+    </MainLayout>
+  );
+}
