+++ app/[lang]/privacy-policy/page.tsx
@@ -1,12 +1,14 @@
 import { setRequestLocale } from 'next-intl/server';
 import React from 'react';
 
+import IntroSection from '~/components/blocks/privacy-policy/intro-section/IntroSection';
+import PolicySection from '~/components/blocks/privacy-policy/policy-section/PolicySection';
+
 import { Language } from '~/types/types/language';
+import { createSeoMeta } from '~/utils/createSeoMeta';
 
 import { createRequestContainer } from '~/di/container';
-import { createSeoMeta } from '~/lib/utils/createSeoMeta';
-import IntroSection from '~/shared/components/blocks/privacy-policy/intro-section/IntroSection';
-import PolicySection from '~/shared/components/blocks/privacy-policy/policy-section/PolicySection';
+import MainLayout from '~/layouts/main-layout/MainLayout';
 
 export const metadata = createSeoMeta({
   title: 'Політика Конфіденційності',
@@ -24,7 +26,7 @@ export default async function PrivacyPolicy({ params }: Readonly<Language>) {
   const blocks = page?.blocks ?? {};
 
   return (
-    <>
+    <MainLayout withLines>
       {blocks.IntroSection && (
         <IntroSection
           title={page.title}
@@ -92,6 +94,6 @@ export default async function PrivacyPolicy({ params }: Readonly<Language>) {
       )}
 
       {blocks.ContactUs && <PolicySection title={blocks.ContactUs.title} description={blocks.ContactUs.description} />}
-    </>
+    </MainLayout>
   );
 }
